Move errors to http package & update tests
