+++ sync/http/errors_test.go
@@ -1,4 +1,4 @@
-package sync
+package http
 
 import (
 	"testing"
